+++ src/components/side-panel.tsx
@@ -91,9 +91,7 @@ export function SidePanel({
 							</div>
 
 							{/* Panel Body - Scrollable */}
-							<div className='flex-1 min-h-0 flex flex-col overflow-y-auto'>
-								{children}
-							</div>
+							<div className='flex-1 min-h-0 flex flex-col p-4'>{children}</div>
 
 							{footer && (
 								<div className='flex h-fit border-t border-zinc-800 p-4'>